Update: tensorboard to version 1.15.0
tensorboard is a runtime dependency for tensorflow 1.15.0

In this version buildcache-1.15.0.tar.gz was updated to
properly populate /tmp/cache

A new tarball is introduced to let bazel manage dependencies
yarn_install provides npm to tensorboard using yarn.  Although
yarn is not able to recognize /tmp/cache hits so the full
.cache/yarn/v4 folder is provided.

Using these tarballs substitute the need for InstallCache.

Modified the pip_package run command to reflect changes on
the new version

Signed-off-by: Daniel Charles <daniel.charles@intel.com>