Fix compatibility with bazel 3.7.0
The `--record_rule_instantiation_callstack` option is no longer
recognized for `bazel build` or `bazel run`. Based on evidence from
upstream commits, it seems that the option was meant for the various
query commands, not for the build command.

Also rev the bazel_java_tools version to reflect an updated dependency
of bazel.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>